DIRECTOR: Track timeout figures as signed integer
Previously, using unsigned integers made it possible to overflow the
timeout check, as you can redefine the tick baseline (and therefore
the last timeout figure) via Lingo.

Fixes catching mice in The Great Green Mouse Disaster.

+++ engines/director/events.cpp
@@ -38,7 +38,7 @@
 
 namespace Director {
 
-uint32 DirectorEngine::getMacTicks() { return (g_system->getMillis() * 60 / 1000.) - _tickBaseline; }
+int DirectorEngine::getMacTicks() { return (int)(g_system->getMillis() * 60 / 1000.) - _tickBaseline; }
 
 bool DirectorEngine::pollEvent(Common::Event &event) {
 	// used by UnitTest XObject
